requirements.txt
azure-functions
openai>=1.35.7
azure-search-documents>=11.6.0b8
httpx>=0.27.0
tiktoken>=0.7.0
langchain>=0.2.10
langchain-openai>=0.1.17
